refactor: reader is not a util
