+++ src/staging/reader.rs
@@ -14,17 +14,148 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
+ *
  */
 
 use std::{
+    fs::{remove_file, File},
     io::{self, BufReader, Read, Seek, SeekFrom},
+    path::PathBuf,
+    sync::Arc,
     vec::IntoIter,
 };
 
-use arrow_array::{RecordBatch, UInt64Array};
+use arrow_array::{RecordBatch, TimestampMillisecondArray};
 use arrow_ipc::{reader::StreamReader, root_as_message_unchecked, MessageHeader};
-use arrow_select::take::take;
+use arrow_schema::Schema;
 use byteorder::{LittleEndian, ReadBytesExt};
+use itertools::kmerge_by;
+use tracing::error;
+
+use crate::{
+    event::DEFAULT_TIMESTAMP_KEY,
+    utils::arrow::{adapt_batch, reverse},
+};
+
+#[derive(Debug)]
+pub struct MergedRecordReader {
+    pub readers: Vec<StreamReader<BufReader<File>>>,
+}
+
+impl MergedRecordReader {
+    pub fn try_new(files: &[PathBuf]) -> Result<Self, ()> {
+        let mut readers = Vec::with_capacity(files.len());
+
+        for file in files {
+            //remove empty files before reading
+            if file.metadata().unwrap().len() == 0 {
+                error!("Invalid file detected, removing it: {:?}", file);
+                remove_file(file).unwrap();
+            } else {
+                let Ok(reader) =
+                    StreamReader::try_new(BufReader::new(File::open(file).unwrap()), None)
+                else {
+                    error!("Invalid file detected, ignoring it: {:?}", file);
+                    continue;
+                };
+
+                readers.push(reader);
+            }
+        }
+
+        Ok(Self { readers })
+    }
+
+    pub fn merged_schema(&self) -> Schema {
+        Schema::try_merge(
+            self.readers
+                .iter()
+                .map(|reader| reader.schema().as_ref().clone()),
+        )
+        .unwrap()
+    }
+}
+
+#[derive(Debug)]
+pub struct MergedReverseRecordReader {
+    pub readers: Vec<StreamReader<BufReader<OffsetReader<File>>>>,
+}
+
+impl MergedReverseRecordReader {
+    pub fn try_new(files: &[PathBuf]) -> Self {
+        let mut readers = Vec::with_capacity(files.len());
+        for file in files {
+            let Ok(reader) = get_reverse_reader(File::open(file).unwrap()) else {
+                error!("Invalid file detected, ignoring it: {:?}", file);
+                continue;
+            };
+
+            readers.push(reader);
+        }
+
+        Self { readers }
+    }
+
+    pub fn merged_iter(
+        self,
+        schema: Arc<Schema>,
+        time_partition: Option<String>,
+    ) -> impl Iterator<Item = RecordBatch> {
+        let adapted_readers = self.readers.into_iter().map(|reader| reader.flatten());
+        kmerge_by(adapted_readers, move |a: &RecordBatch, b: &RecordBatch| {
+            // Capture time_partition by value
+            let a_time = get_timestamp_millis(a, time_partition.clone());
+            let b_time = get_timestamp_millis(b, time_partition.clone());
+            a_time > b_time
+        })
+        .map(|batch| reverse(&batch))
+        .map(move |batch| adapt_batch(&schema, &batch))
+    }
+
+    pub fn merged_schema(&self) -> Schema {
+        Schema::try_merge(
+            self.readers
+                .iter()
+                .map(|reader| reader.schema().as_ref().clone()),
+        )
+        .unwrap()
+    }
+}
+
+fn get_timestamp_millis(batch: &RecordBatch, time_partition: Option<String>) -> i64 {
+    match time_partition {
+        Some(time_partition) => {
+            let time_partition = time_partition.as_str();
+            match batch.column_by_name(time_partition) {
+                Some(column) => column
+                    .as_any()
+                    .downcast_ref::<TimestampMillisecondArray>()
+                    .unwrap()
+                    .value(0),
+                None => get_default_timestamp_millis(batch),
+            }
+        }
+        None => get_default_timestamp_millis(batch),
+    }
+}
+fn get_default_timestamp_millis(batch: &RecordBatch) -> i64 {
+    match batch
+        .column(0)
+        .as_any()
+        .downcast_ref::<TimestampMillisecondArray>()
+    {
+        // Ideally we expect the first column to be a timestamp (because we add the timestamp column first in the writer)
+        Some(array) => array.value(0),
+        // In case the first column is not a timestamp, we fallback to look for default timestamp column across all columns
+        None => batch
+            .column_by_name(DEFAULT_TIMESTAMP_KEY)
+            .unwrap()
+            .as_any()
+            .downcast_ref::<TimestampMillisecondArray>()
+            .unwrap()
+            .value(0),
+    }
+}
 
 /// OffsetReader takes in a reader and list of offset and sizes and
 /// provides a reader over the file by reading only the offsets
@@ -140,16 +271,6 @@ pub fn get_reverse_reader<T: Read + Seek>(
     reader.rewind()?;
 
     Ok(StreamReader::try_new(BufReader::new(OffsetReader::new(reader, messages)), None).unwrap())
-}
-
-pub fn reverse(rb: &RecordBatch) -> RecordBatch {
-    let indices = UInt64Array::from_iter_values((0..rb.num_rows()).rev().map(|x| x as u64));
-    let arrays = rb
-        .columns()
-        .iter()
-        .map(|col| take(&col, &indices, None).unwrap())
-        .collect();
-    RecordBatch::try_new(rb.schema(), arrays).unwrap()
 }
 
 // return limit for
